requirements.txt
Django==5.1.1
psycopg[binary]==3.2.1
django-redis==5.4.0
django-filter==24.3
Pillow==10.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
celery==5.4.0
redis==5.0.7
django-parler==2.3
sentry-sdk==2.12.0
